+++ src/operations/Diagnostics.cpp
@@ -20,7 +20,7 @@ lsp::DocumentDiagnosticReport WorkspaceFolder::documentDiagnostics(const lsp::Do
 
     // TODO: should we apply a resultId and return an unchanged report if unchanged?
     lsp::DocumentDiagnosticReport report;
-    std::unordered_map<std::string /* lsp::DocumentUri */, std::vector<lsp::Diagnostic>> relatedDiagnostics{};
+    std::unordered_map<Uri, std::vector<lsp::Diagnostic>, UriHash> relatedDiagnostics{};
 
     auto moduleName = fileResolver.getModuleName(params.textDocument.uri);
     auto textDocument = fileResolver.getTextDocument(params.textDocument.uri);
@@ -58,12 +58,14 @@ lsp::DocumentDiagnosticReport WorkspaceFolder::documentDiagnostics(const lsp::Do
         else if (supportsRelatedDocuments(client->capabilities))
         {
             auto fileName = platform->resolveToRealPath(error.moduleName);
-            if (!fileName || isIgnoredFile(*fileName, config))
+            if (!fileName)
                 continue;
             auto textDocument = fileResolver.getTextDocumentFromModuleName(error.moduleName);
-            auto diagnostic = createTypeErrorDiagnostic(error, &fileResolver, textDocument);
             auto uri = textDocument ? textDocument->uri() : Uri::file(*fileName);
-            auto& currentDiagnostics = relatedDiagnostics[uri.toString()];
+            if (isIgnoredFile(uri, config))
+                continue;
+            auto diagnostic = createTypeErrorDiagnostic(error, &fileResolver, textDocument);
+            auto& currentDiagnostics = relatedDiagnostics[uri];
             currentDiagnostics.emplace_back(diagnostic);
         }
     }
@@ -216,7 +218,7 @@ void WorkspaceFolder::pushDiagnostics(const lsp::DocumentUri& uri, const size_t 
             {
                 if (relatedDiagnostics.kind == lsp::DocumentDiagnosticReportKind::Full)
                 {
-                    client->publishDiagnostics(lsp::PublishDiagnosticsParams{Uri::parse(relatedUri), std::nullopt, relatedDiagnostics.items});
+                    client->publishDiagnostics(lsp::PublishDiagnosticsParams{relatedUri, std::nullopt, relatedDiagnostics.items});
                 }
             }
         }